Fixes a bug which would cause the app content to be missing in an app if the app was migrated from one Storm App ID to another under a version bump.
This is caused because if `setupRootWindow` is called before
`ContentController.shared.appLaunched()` is called then the
app.json is loaded from the previous app's delta bundle , which
is then subsequently deleted by the content controller when it
detects an.app version change.

This should all be okay and not result in any issues, due to these
calls being synchronous, but it seems that there's some unforseen
asynchronous process which means that the files are deleted by
the content controller before the content is shown on-screen.

Previously this would have only resulted in the user being shown the
same content as the latest delta update for the app in question,
but due some underlying issue it was resulting in a blank app until
next launch at which point the previous App ID's delta is fully removed